requirements.txt
python-dotenv

tqdm
torch
numpy

sqlite3
requests
beautifulsoup4

pillow
pytesseract
konlpy

rank-bm25
transformers

streamlit
openai
